vergeScraper test mode, connection try/except,
Added a test mode to the verge scraper script. Running the scraper with the following command shortens the list of phones scraped and written to 8 for testing purposes. 
python vergeScraper.py dbTest 

Also added a function which recursively tries to connect to the database. If it fails 5 times, it exits reporting the  exception.  

Removed redundant # in the Pro and Con model class string methods.