Python Module interfacing with PriceAPI
Adds the priceApiInterface.py file to the CellCheck directory. This provides functions for querying the PriceAPI, and recieving data in a structure of nested dictionaries and lists. This structure can easily be turned into a JSON object for use in AJAX. 

a minified version of jquery was also added to the static files for use in front-end javascript.